+++ scripts/render-details4.sh
@@ -423,10 +423,12 @@ validate_jsonld() {
     REPORTFILE=${RLINE}/jsonld-validation.report.md
     echo "${REPORTLINEPREFIX}oslo-jsonld-validator ${REPORTLINENEWLINE}" &>>${REPORTFILE}
     echo "${REPORTLINEPREFIX}-------------------------------------${REPORTLINENEWLINE}" &>>${REPORTFILE}
+    echo "$URIDOMAIN"
     
     oslo-jsonld-validator --input ${MERGEDFILE} \
     --whitelist https://raw.githubusercontent.com/Informatievlaanderen/OSLO-UML-Transformer/refs/heads/configuration/whitelist.json \
     --specificationType ${SPECTYPE} \
+    --publicationEnvironment $URIDOMAIN \
     2>&1 | tee -a ${REPORTFILE}
     
     echo ${REPORTFILE}
@@ -1153,7 +1155,7 @@ render_swagger() {
             cat ${REPORTFILE}
             execution_strickness
         fi
-        if [ -f ${TLINE} ]; then
+        if [ -d ${TLINE} ]; then
             echo "RENDER-DETAILs: success"
         else
             echo "RENDER-DETAILS: failed"