+++ CHANGELOG.md
@@ -145,3 +145,4 @@ Upgrading from version 3.x.y to 4.x.y is thus a non-trivial task.
 - Add bundling to the publication point configuration to allow for bundling of publications. This will fetch all potential assets like codelists, external ontologies, etc.
 - improvements on bundling process
 - Bumped informatievlaanderen/oslo-specification-generator:multilingual-dev4.0 to informatievlaanderen/oslo-specification-generator:multilingual-dev4.0.3
+- Improved templates